NEMS commits to add GAEA and Jet support.  Various bug fixes are added to NEMSfv3gfs.
1. GAEA support is added to NEMSCompsetRun and NEMSAppBuilder.  The
NEMSfv3gfs compsets all work on GAEA C3 and C4.

2. Jet IMPI support in NEMSfv3gfs, plus support for C768 on Jet.  Both
xJet and the s+v Jet pair are supported and have separate baselines.

3. Pass $CFLAGS down to NEMS C code so we can add uJet and tJet
support.  It will be a while before the underlying libraries are
recompiled; until then we cannot support anything on t or u Jet.

4. Add FV3-MOM6-CICE5 to the list of Supported apps.

5. On sJet and vJet, the C768 no longer fits in 16 cores per node.  I
have reduced it to 14 cores per node.

6. On WCOSS Phase 2, some of the tests now take slightly over a half
hour (sometimes).  I reverted the queue to dev2 and wallclock limit to
45 minutes.  If you feel like baby-sitting the workflow, you can
switch back to debug2 and 30 minutes, but you will have to keep
running "rtrewind" on the jobs that hit their wallclock limit.
Ultimately, you might finish the test faster that way, but it will be
quite tedious.