Replace the NEMSAppBuilder with a new system based on GNU Make. Update the documentation.
The main entry point to the build system is in NEMS/GNUmakefile, which
contains comments with an overview of the internal logic.  The
per-platform logic is within NEMS/src/incmake/env, and the
per-component build rules are in NEMS/src/incmake/component_FOO.mk for
each component FOO.

The NEMSAppBuilder is now a simple wrapper around the
NEMS/GNUmakefile.  The NEMSfv3gfs/tests/compile.sh is updated for the
new makefiles.

An FV3 change is needed for this new build system, to fix a bug in the
FV3 top-level makefile wherein Make's special MAKEFLAGS variable is
misused to pass down internal build options.  This breaks the build if
certain make flags are used, which are not preceeded by a "-" when
stored in the MAKEFLAGS.